Update to 1.19
This builds. Lets just hope i did not miss any bugs...
Fixes #39

+++ src/main/java/me/b0iizz/advancednbttooltip/api/impl/builtin/FuelTimeFactory.java
@@ -25,7 +25,6 @@ package me.b0iizz.advancednbttooltip.api.impl.builtin;
 import me.b0iizz.advancednbttooltip.api.JsonTooltips;
 import me.b0iizz.advancednbttooltip.api.TooltipFactory;
 import net.fabricmc.fabric.api.registry.FuelRegistry;
-import net.minecraft.block.ComposterBlock;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.Item;
 import net.minecraft.nbt.NbtCompound;
@@ -34,7 +33,6 @@ import net.minecraft.text.Text;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalInt;
 
 /**
  * @author B0IIZZ
@@ -44,7 +42,7 @@ public class FuelTimeFactory implements TooltipFactory {
 
 	@Override
 	public List<Text> getTooltipText(Item item, NbtCompound tag, TooltipContext context) {
-		int time = Optional.ofNullable(FuelRegistry.INSTANCE.get(item)).orElse(0).intValue();
+		int time = Optional.ofNullable(FuelRegistry.INSTANCE.get(item)).orElse(0);
 		return time <= 0 ? Collections.emptyList() : List.of(Text.of(String.valueOf(time)));
 	}
 }
